+++ crates/iota-sdk/src/wallet_context.rs
@@ -14,13 +14,13 @@ use iota_json_rpc_types::{
     IotaTransactionBlockResponse, IotaTransactionBlockResponseOptions,
 };
 use iota_keys::keystore::{AccountKeystore, Keystore};
+use iota_sdk_types::crypto::Intent;
 use iota_types::{
     base_types::{IotaAddress, ObjectID, ObjectRef},
     crypto::IotaKeyPair,
     gas_coin::GasCoin,
     transaction::{Transaction, TransactionData, TransactionDataAPI},
 };
-use shared_crypto::intent::Intent;
 use tokio::sync::RwLock;
 use tracing::warn;