+++ crates/starfish/core/src/authority_node.rs
@@ -181,7 +181,6 @@ impl ConsensusAuthority {
             network_client.clone(),
             context.clone(),
             core_dispatcher.clone(),
-            block_verifier.clone(),
             dag_state.clone(),
         );
 
@@ -212,7 +211,6 @@ impl ConsensusAuthority {
             context.clone(),
             core_dispatcher.clone(),
             commit_vote_monitor.clone(),
-            transactions_synchronizer.clone(),
             commit_consumer_monitor.clone(),
             network_client.clone(),
             block_verifier.clone(),