+++ docs/examples/rust/src/utils.rs
@@ -22,7 +22,7 @@ use iota_sdk::{
         transaction::{Transaction, TransactionData},
     },
 };
-use shared_crypto::intent::Intent;
+use iota_sdk_types::crypto::Intent;
 
 /// Got from iota-genesis-builder/src/stardust/test_outputs/stardust_mix.rs
 const SPONSOR_ADDRESS_MNEMONIC: &str = "okay pottery arch air egg very cave cash poem gown sorry mind poem crack dawn wet car pink extra crane hen bar boring salt";