+++ sdk/dapp-kit/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @iota/dapp-kit
 
+## 0.7.0
+
+### Minor Changes
+
+-   147c97e: Update repository URLs.
+
+### Patch Changes
+
+-   55fab22: Add onConnected callback to connect button
+-   Updated dependencies [00f4a39]
+-   Updated dependencies [f4cc8e8]
+-   Updated dependencies [933496c]
+-   Updated dependencies [0a4525d]
+-   Updated dependencies [ec99569]
+    -   @iota/iota-sdk@1.8.0
+    -   @iota/wallet-standard@0.2.17
+
 ## 0.6.0
 
 ### Minor Changes